use int rather than np.int in cython

+++ landlab/components/flow_director/cfuncs.pyx
@@ -6,7 +6,7 @@ cimport cython
 DTYPE_FLOAT = np.double
 ctypedef np.double_t DTYPE_FLOAT_t
 
-DTYPE_INT = np.int
+DTYPE_INT = int
 #ctypedef np.longlong_t DTYPE_INT_t
 ctypedef np.int_t DTYPE_INT_t
